Refactor implementation
Strategy pattern for varying storage saving mechanism
Singleton pattern for in memory storage
Factory pattern for storage mechanism retrieval
Business logic layer enhancements
TODO: lot of stuff left

+++ src/Transaction.java
@@ -4,15 +4,15 @@ import java.util.Date;
  * Created by Ovidiu on 10/14/2017.
  */
 public class Transaction {
-    private int sum;
+    private double sum;
     private Date date;
 
-    public int getSum()
+    public double getSum()
     {
         return sum;
     }
 
-    public void setSum(int sum)
+    public void setSum(double sum)
     {
         this.sum = sum;
     }
